oop_edu_back/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
PyMySQL==1.1.0
redis==5.0.1
qiniu==7.10.0
openai==1.3.0
jieba==0.42.1
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
email-validator==2.1.0